+++ packages/next/lib/load-custom-routes.ts
@@ -3,9 +3,15 @@ import type { Token } from 'next/dist/compiled/path-to-regexp'
 
 import chalk from './chalk'
 import { escapeStringRegexp } from '../shared/lib/escape-regexp'
-import { PERMANENT_REDIRECT_STATUS } from '../shared/lib/constants'
-import { TEMPORARY_REDIRECT_STATUS } from '../shared/lib/constants'
 import { tryToParsePath } from './try-to-parse-path'
+
+import {
+  getRedirectStatus,
+  allowedStatusCodes,
+  modifyRouteRegex,
+} from './get-redirect-status'
+
+export { getRedirectStatus, allowedStatusCodes, modifyRouteRegex }
 
 export type RouteHas =
   | {
@@ -46,39 +52,12 @@ export type Redirect = {
   permanent?: boolean
 }
 
-export const allowedStatusCodes = new Set([301, 302, 303, 307, 308])
 const allowedHasTypes = new Set(['header', 'cookie', 'query', 'host'])
 const namedGroupsRegex = /\(\?<([a-zA-Z][a-zA-Z0-9]*)>/g
-
-export function getRedirectStatus(route: {
-  statusCode?: number
-  permanent?: boolean
-}): number {
-  return (
-    route.statusCode ||
-    (route.permanent ? PERMANENT_REDIRECT_STATUS : TEMPORARY_REDIRECT_STATUS)
-  )
-}
 
 export function normalizeRouteRegex(regex: string) {
   // clean up un-necessary escaping from regex.source which turns / into \\/
   return regex.replace(/\\\//g, '/')
-}
-
-// for redirects we restrict matching /_next and for all routes
-// we add an optional trailing slash at the end for easier
-// configuring between trailingSlash: true/false
-export function modifyRouteRegex(regex: string, restrictedPaths?: string[]) {
-  if (restrictedPaths) {
-    regex = regex.replace(
-      /\^/,
-      `^(?!${restrictedPaths
-        .map((path) => path.replace(/\//g, '\\/'))
-        .join('|')})`
-    )
-  }
-  regex = regex.replace(/\$$/, '(?:\\/)?$')
-  return regex
 }
 
 function checkRedirect(route: Redirect): {
