+++ packages/next/server/base-server.ts
@@ -1,4 +1,4 @@
-import { __ApiPreviewProps } from './api-utils'
+import type { __ApiPreviewProps } from './api-utils'
 import type { CustomRoutes } from '../lib/load-custom-routes'
 import type { DomainLocale } from './config'
 import type { DynamicRoutes, PageChecker, Route } from './router'
@@ -20,6 +20,12 @@ import type {
   ResponseCacheValue,
 } from './response-cache'
 import type { UrlWithParsedQuery } from 'url'
+import type { PreviewData } from 'next/types'
+import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
+import type { BaseNextRequest, BaseNextResponse } from './base-http'
+import type { PayloadOptions } from './send-payload'
+import type { PrerenderManifest } from '../build'
+
 import {
   CacheFs,
   NormalizeError,
@@ -27,15 +33,10 @@ import {
   normalizeRepeatedSlashes,
   MissingStaticPage,
 } from '../shared/lib/utils'
-import type { PreviewData } from 'next/types'
-import type { PagesManifest } from '../build/webpack/plugins/pages-manifest-plugin'
-import type { BaseNextRequest, BaseNextResponse } from './base-http'
-import type { PayloadOptions } from './send-payload'
-
 import { join, resolve } from '../shared/lib/isomorphic/path'
 import { parse as parseQs } from 'querystring'
 import { format as formatUrl, parse as parseUrl } from 'url'
-import { getRedirectStatus } from '../lib/load-custom-routes'
+import { getRedirectStatus } from '../lib/get-redirect-status'
 import {
   NEXT_BUILTIN_DOCUMENT,
   NEXT_CLIENT_SSR_ENTRY_SUFFIX,
@@ -68,8 +69,6 @@ import escapePathDelimiters from '../shared/lib/router/utils/escape-path-delimit
 import { getUtils } from '../build/webpack/loaders/next-serverless-loader/utils'
 import isError, { getProperError } from '../lib/is-error'
 import { addRequestMeta, getRequestMeta } from './request-meta'
-import { createHeaderRoute, createRedirectRoute } from './server-route-utils'
-import { PrerenderManifest } from '../build'
 import { ImageConfigComplete } from '../shared/lib/image-config'
 import { removePathPrefix } from '../shared/lib/router/utils/remove-path-prefix'
 import { normalizeAppPath } from '../shared/lib/router/utils/app-paths'
@@ -202,11 +201,6 @@ export default abstract class Server<ServerOptions extends Options = Options> {
   protected abstract getPagesManifest(): PagesManifest | undefined
   protected abstract getAppPathsManifest(): PagesManifest | undefined
   protected abstract getBuildId(): string
-  protected abstract generatePublicRoutes(): Route[]
-  protected abstract generateImageRoutes(): Route[]
-  protected abstract generateStaticRoutes(): Route[]
-  protected abstract generateFsStaticRoutes(): Route[]
-  protected abstract generateCatchAllMiddlewareRoute(): Route[]
   protected abstract generateRewrites({
     restrictedRedirectPaths,
   }: {
@@ -223,7 +217,6 @@ export default abstract class Server<ServerOptions extends Options = Options> {
     params?: Params,
     isAppDir?: boolean
   ): Promise<FindComponentsResult | null>
-  protected abstract getPagePath(pathname: string, locales?: string[]): string
   protected abstract getFontManifest(): FontManifest | undefined
   protected abstract getRoutesManifest(): CustomRoutes
   protected abstract getPrerenderManifest(): PrerenderManifest
@@ -232,6 +225,24 @@ export default abstract class Server<ServerOptions extends Options = Options> {
     req: BaseNextRequest,
     parsedUrl: NextUrlWithParsedQuery
   ): void
+  protected abstract hasPage(pathname: string): Promise<boolean>
+
+  protected abstract generateRoutes(): {
+    headers: Route[]
+    rewrites: {
+      beforeFiles: Route[]
+      afterFiles: Route[]
+      fallback: Route[]
+    }
+    fsRoutes: Route[]
+    redirects: Route[]
+    catchAllRoute: Route
+    catchAllMiddleware: Route[]
+    pageChecker: PageChecker
+    useFileSystemPublicRoutes: boolean
+    dynamicRoutes: DynamicRoutes | undefined
+    nextConfig: NextConfig
+  }
 
   protected abstract sendRenderResult(
     req: BaseNextRequest,
@@ -691,227 +702,87 @@ export default abstract class Server<ServerOptions extends Options = Options> {
     return this.getPrerenderManifest().preview
   }
 
-  protected generateRoutes(): {
-    headers: Route[]
-    rewrites: {
-      beforeFiles: Route[]
-      afterFiles: Route[]
-      fallback: Route[]
-    }
-    fsRoutes: Route[]
-    redirects: Route[]
-    catchAllRoute: Route
-    catchAllMiddleware: Route[]
-    pageChecker: PageChecker
-    useFileSystemPublicRoutes: boolean
-    dynamicRoutes: DynamicRoutes | undefined
-    nextConfig: NextConfig
-  } {
-    const publicRoutes = this.generatePublicRoutes()
-    const imageRoutes = this.generateImageRoutes()
-    const staticFilesRoutes = this.generateStaticRoutes()
-
-    const fsRoutes: Route[] = [
-      ...this.generateFsStaticRoutes(),
-      {
-        match: getPathMatch('/_next/data/:path*'),
-        type: 'route',
-        name: '_next/data catchall',
-        fn: async (req, res, params, _parsedUrl) => {
-          // Make sure to 404 for /_next/data/ itself and
-          // we also want to 404 if the buildId isn't correct
-          if (!params.path || params.path[0] !== this.buildId) {
-            await this.render404(req, res, _parsedUrl)
-            return {
-              finished: true,
-            }
-          }
-          // remove buildId from URL
-          params.path.shift()
-
-          const lastParam = params.path[params.path.length - 1]
-
-          // show 404 if it doesn't end with .json
-          if (typeof lastParam !== 'string' || !lastParam.endsWith('.json')) {
-            await this.render404(req, res, _parsedUrl)
-            return {
-              finished: true,
-            }
-          }
-
-          // re-create page's pathname
-          let pathname = `/${params.path.join('/')}`
-          pathname = getRouteFromAssetPath(pathname, '.json')
-
-          // ensure trailing slash is normalized per config
-          if (this.router.catchAllMiddleware[0]) {
-            if (this.nextConfig.trailingSlash && !pathname.endsWith('/')) {
-              pathname += '/'
-            }
-            if (
-              !this.nextConfig.trailingSlash &&
-              pathname.length > 1 &&
-              pathname.endsWith('/')
-            ) {
-              pathname = pathname.substring(0, pathname.length - 1)
-            }
-          }
-
-          if (this.nextConfig.i18n) {
-            const { host } = req?.headers || {}
-            // remove port from host and remove port if present
-            const hostname = host?.split(':')[0].toLowerCase()
-            const localePathResult = normalizeLocalePath(
-              pathname,
-              this.nextConfig.i18n.locales
-            )
-            const { defaultLocale } =
-              detectDomainLocale(this.nextConfig.i18n.domains, hostname) || {}
-
-            let detectedLocale = ''
-
-            if (localePathResult.detectedLocale) {
-              pathname = localePathResult.pathname
-              detectedLocale = localePathResult.detectedLocale
-            }
-
-            _parsedUrl.query.__nextLocale = detectedLocale
-            _parsedUrl.query.__nextDefaultLocale =
-              defaultLocale || this.nextConfig.i18n.defaultLocale
-
-            if (!detectedLocale && !this.router.catchAllMiddleware[0]) {
-              _parsedUrl.query.__nextLocale =
-                _parsedUrl.query.__nextDefaultLocale
-              await this.render404(req, res, _parsedUrl)
-              return { finished: true }
-            }
-          }
-
-          return {
-            pathname,
-            query: { ..._parsedUrl.query, __nextDataReq: '1' },
-            finished: false,
-          }
-        },
-      },
-      ...imageRoutes,
-      {
-        match: getPathMatch('/_next/:path*'),
-        type: 'route',
-        name: '_next catchall',
-        // This path is needed because `render()` does a check for `/_next` and the calls the routing again
-        fn: async (req, res, _params, parsedUrl) => {
-          await this.render404(req, res, parsedUrl)
+  protected generateDataCatchallRoute(): Route {
+    return {
+      match: getPathMatch('/_next/data/:path*'),
+      type: 'route',
+      name: '_next/data catchall',
+      fn: async (req, res, params, _parsedUrl) => {
+        // Make sure to 404 for /_next/data/ itself and
+        // we also want to 404 if the buildId isn't correct
+        if (!params.path || params.path[0] !== this.buildId) {
+          await this.render404(req, res, _parsedUrl)
           return {
             finished: true,
           }
-        },
-      },
-      ...publicRoutes,
-      ...staticFilesRoutes,
-    ]
+        }
+        // remove buildId from URL
+        params.path.shift()
 
-    const restrictedRedirectPaths = this.nextConfig.basePath
-      ? [`${this.nextConfig.basePath}/_next`]
-      : ['/_next']
+        const lastParam = params.path[params.path.length - 1]
 
-    // Headers come very first
-    const headers = this.minimalMode
-      ? []
-      : this.customRoutes.headers.map((rule) =>
-          createHeaderRoute({ rule, restrictedRedirectPaths })
-        )
-
-    const redirects = this.minimalMode
-      ? []
-      : this.customRoutes.redirects.map((rule) =>
-          createRedirectRoute({ rule, restrictedRedirectPaths })
-        )
-
-    const rewrites = this.generateRewrites({ restrictedRedirectPaths })
-    const catchAllMiddleware = this.generateCatchAllMiddlewareRoute()
-
-    const catchAllRoute: Route = {
-      match: getPathMatch('/:path*'),
-      type: 'route',
-      matchesLocale: true,
-      name: 'Catchall render',
-      fn: async (req, res, _params, parsedUrl) => {
-        let { pathname, query } = parsedUrl
-        if (!pathname) {
-          throw new Error('pathname is undefined')
+        // show 404 if it doesn't end with .json
+        if (typeof lastParam !== 'string' || !lastParam.endsWith('.json')) {
+          await this.render404(req, res, _parsedUrl)
+          return {
+            finished: true,
+          }
         }
 
-        // next.js core assumes page path without trailing slash
-        pathname = removeTrailingSlash(pathname)
+        // re-create page's pathname
+        let pathname = `/${params.path.join('/')}`
+        pathname = getRouteFromAssetPath(pathname, '.json')
+
+        // ensure trailing slash is normalized per config
+        if (this.router.catchAllMiddleware[0]) {
+          if (this.nextConfig.trailingSlash && !pathname.endsWith('/')) {
+            pathname += '/'
+          }
+          if (
+            !this.nextConfig.trailingSlash &&
+            pathname.length > 1 &&
+            pathname.endsWith('/')
+          ) {
+            pathname = pathname.substring(0, pathname.length - 1)
+          }
+        }
 
         if (this.nextConfig.i18n) {
+          const { host } = req?.headers || {}
+          // remove port from host and remove port if present
+          const hostname = host?.split(':')[0].toLowerCase()
           const localePathResult = normalizeLocalePath(
             pathname,
-            this.nextConfig.i18n?.locales
+            this.nextConfig.i18n.locales
           )
+          const { defaultLocale } =
+            detectDomainLocale(this.nextConfig.i18n.domains, hostname) || {}
+
+          let detectedLocale = ''
 
           if (localePathResult.detectedLocale) {
             pathname = localePathResult.pathname
-            parsedUrl.query.__nextLocale = localePathResult.detectedLocale
+            detectedLocale = localePathResult.detectedLocale
           }
-        }
-        const bubbleNoFallback = !!query._nextBubbleNoFallback
 
-        if (pathname === '/api' || pathname.startsWith('/api/')) {
-          delete query._nextBubbleNoFallback
+          _parsedUrl.query.__nextLocale = detectedLocale
+          _parsedUrl.query.__nextDefaultLocale =
+            defaultLocale || this.nextConfig.i18n.defaultLocale
 
-          const handled = await this.handleApiRequest(req, res, pathname, query)
-          if (handled) {
+          if (!detectedLocale && !this.router.catchAllMiddleware[0]) {
+            _parsedUrl.query.__nextLocale = _parsedUrl.query.__nextDefaultLocale
+            await this.render404(req, res, _parsedUrl)
             return { finished: true }
           }
         }
 
-        try {
-          await this.render(req, res, pathname, query, parsedUrl, true)
-
-          return {
-            finished: true,
-          }
-        } catch (err) {
-          if (err instanceof NoFallbackError && bubbleNoFallback) {
-            return {
-              finished: false,
-            }
-          }
-          throw err
+        return {
+          pathname,
+          query: { ..._parsedUrl.query, __nextDataReq: '1' },
+          finished: false,
         }
       },
     }
-
-    const { useFileSystemPublicRoutes } = this.nextConfig
-
-    if (useFileSystemPublicRoutes) {
-      this.appPathRoutes = this.getAppPathRoutes()
-      this.dynamicRoutes = this.getDynamicRoutes()
-    }
-
-    return {
-      headers,
-      fsRoutes,
-      rewrites,
-      redirects,
-      catchAllRoute,
-      catchAllMiddleware,
-      useFileSystemPublicRoutes,
-      dynamicRoutes: this.dynamicRoutes,
-      pageChecker: this.hasPage.bind(this),
-      nextConfig: this.nextConfig,
-    }
-  }
-
-  protected async hasPage(pathname: string): Promise<boolean> {
-    let found = false
-    try {
-      found = !!this.getPagePath(pathname, this.nextConfig.i18n?.locales)
-    } catch (_) {}
-
-    return found
   }
 
   protected async _beforeCatchAllRender(
@@ -923,55 +794,12 @@ export default abstract class Server<ServerOptions extends Options = Options> {
     return false
   }
 
-  // Used to build API page in development
-  protected async ensureApiPage(_pathname: string): Promise<void> {}
-
-  /**
-   * Resolves `API` request, in development builds on demand
-   * @param req http request
-   * @param res http response
-   * @param pathname path of request
-   */
-  private async handleApiRequest(
+  protected abstract handleApiRequest(
     req: BaseNextRequest,
     res: BaseNextResponse,
     pathname: string,
     query: ParsedUrlQuery
-  ): Promise<boolean> {
-    let page = pathname
-    let params: Params | undefined = undefined
-    let pageFound = !isDynamicRoute(page) && (await this.hasPage(page))
-
-    if (!pageFound && this.dynamicRoutes) {
-      for (const dynamicRoute of this.dynamicRoutes) {
-        params = dynamicRoute.match(pathname) || undefined
-        if (dynamicRoute.page.startsWith('/api') && params) {
-          page = dynamicRoute.page
-          pageFound = true
-          break
-        }
-      }
-    }
-
-    if (!pageFound) {
-      return false
-    }
-    // Make sure the page is built before getting the path
-    // or else it won't be in the manifest yet
-    await this.ensureApiPage(page)
-
-    let builtPagePath
-    try {
-      builtPagePath = this.getPagePath(page)
-    } catch (err) {
-      if (isError(err) && err.code === 'ENOENT') {
-        return false
-      }
-      throw err
-    }
-
-    return this.runApi(req, res, query, params, page, builtPagePath)
-  }
+  ): Promise<boolean>
 
   protected getDynamicRoutes(): Array<RoutingItem> {
     const addedPages = new Set<string>()
@@ -2113,7 +1941,7 @@ export function prepareServerlessUrl(
 
 export { stringifyQuery } from './server-route-utils'
 
-class NoFallbackError extends Error {}
+export class NoFallbackError extends Error {}
 
 // Internal wrapper around build errors at development
 // time, to prevent us from propagating or logging them
