+++ packages/next/server/web/spec-extension/cookies.ts
@@ -1,5 +1,5 @@
+import type { CookieSerializeOptions } from '../types'
 import cookie from 'next/dist/compiled/cookie'
-import { CookieSerializeOptions } from '../types'
 
 type GetWithOptionsOutput = {
   value: string | undefined
